cpp/requirements.txt
jinja2
progressbar33
ruamel_yaml
pycryptodome
python-magic
